fix(web-ui): replace unsafe vcStatus casts across execution components
