+++ control-plane/web/client/src/components/execution/EnhancedExecutionHeader.tsx
@@ -11,6 +11,7 @@ import {
 } from "@/components/ui/icon-bridge";
 import { useNavigate } from "react-router-dom";
 import type { WorkflowExecution } from "../../types/executions";
+import type { VCStatusData } from "../../types/did";
 import { DIDDisplay } from "../did/DIDDisplay";
 import { Button } from "../ui/button";
 import { Badge } from "../ui/badge";
@@ -22,13 +23,7 @@ import { CopyButton } from "../ui/copy-button";
 
 interface EnhancedExecutionHeaderProps {
   execution: WorkflowExecution;
-  vcStatus?: {
-    has_vc: boolean;
-    vc_id?: string;
-    status: string;
-    created_at?: string;
-    vc_document?: any;
-  } | null;
+  vcStatus?: VCStatusData | null;
   vcLoading?: boolean;
   onClose?: () => void;
 }
@@ -326,7 +321,7 @@ export function EnhancedExecutionHeader({
                 <VerifiableCredentialBadge
                   hasVC={vcStatus.has_vc}
                   status={vcStatus.status}
-                  vcData={vcStatus as any}
+                  vcData={vcStatus}
                   executionId={execution.execution_id}
                   showCopyButton={false}
                   showVerifyButton={false}
