+++ control-plane/web/client/src/components/execution/ExecutionHero.tsx
@@ -11,6 +11,7 @@ import {
 } from "@/components/ui/icon-bridge";
 import { useNavigate } from "react-router-dom";
 import type { WorkflowExecution } from "../../types/executions";
+import type { VCStatusData } from "../../types/did";
 import { DIDDisplay } from "../did/DIDDisplay";
 import { Button } from "../ui/button";
 import { Card, CardContent } from "../ui/card";
@@ -23,13 +24,7 @@ import { cn } from "@/lib/utils";
 
 interface ExecutionHeroProps {
   execution: WorkflowExecution;
-  vcStatus?: {
-    has_vc: boolean;
-    vc_id?: string;
-    status: string;
-    created_at?: string;
-    vc_document?: any;
-  } | null;
+  vcStatus?: VCStatusData | null;
   vcLoading?: boolean;
 }
 
@@ -38,13 +33,7 @@ function VCStatusCard({
   vcLoading,
   executionId,
 }: {
-  vcStatus?: {
-    has_vc: boolean;
-    vc_id?: string;
-    status: string;
-    created_at?: string;
-    vc_document?: any;
-  } | null;
+  vcStatus?: VCStatusData | null;
   vcLoading?: boolean;
   executionId: string;
 }) {
@@ -69,7 +58,7 @@ function VCStatusCard({
           <VerifiableCredentialBadge
             hasVC={vcStatus.has_vc}
             status={vcStatus.status}
-            vcData={vcStatus as any}
+            vcData={vcStatus}
             executionId={executionId}
             showCopyButton={true}
             showVerifyButton={true}
@@ -134,7 +123,7 @@ export function ExecutionHero({
                   {execution.reasoner_id}
                 </h1>
                 <StatusIndicator
-                  status={execution.status as any}
+                  status={execution.status}
                   animated={execution.status === "running"}
                   className="text-base"
                 />
