+++ control-plane/web/client/src/components/execution/ExecutionIdentityPanel.tsx
@@ -2,6 +2,7 @@ import { useState } from "react";
 import { Shield, ExternalLink, AlertCircle, CheckCircle, Eye, Download, Loader2 } from "@/components/ui/icon-bridge";
 import { ResponsiveGrid } from "@/components/layout/ResponsiveGrid";
 import type { WorkflowExecution } from "../../types/executions";
+import type { VCStatusData, VCDocument } from "../../types/did";
 import { DIDDisplay } from "../did/DIDDisplay";
 import { Button } from "../ui/button";
 import { Card, CardContent, CardHeader, CardTitle } from "../ui/card";
@@ -13,13 +14,7 @@ import { CopyButton } from "../ui/copy-button";
 
 interface ExecutionIdentityPanelProps {
   execution: WorkflowExecution;
-  vcStatus?: {
-    has_vc: boolean;
-    vc_id?: string;
-    status: string;
-    created_at?: string;
-    vc_document?: any;
-  } | null;
+  vcStatus?: VCStatusData | null;
   vcLoading?: boolean;
 }
 
@@ -56,7 +51,7 @@ export function ExecutionIdentityPanel({
 
   const vcDocument =
     vcStatus?.vc_document && typeof vcStatus.vc_document === "object" && !Array.isArray(vcStatus.vc_document)
-      ? (vcStatus.vc_document as Record<string, any>)
+      ? (vcStatus.vc_document as Partial<VCDocument>)
       : null;
   const credentialSubject = vcDocument?.credentialSubject ?? null;
   const executionDetails = credentialSubject?.execution ?? null;
@@ -166,7 +161,7 @@ export function ExecutionIdentityPanel({
                     <VerifiableCredentialBadge
                       hasVC={vcStatus.has_vc}
                       status={vcStatus.status}
-                      vcData={vcStatus as any}
+                      vcData={vcStatus}
                       executionId={execution.execution_id}
                       showCopyButton={false}
                       showVerifyButton={false}
@@ -371,7 +366,7 @@ export function ExecutionIdentityPanel({
                   </div>
                 </ResponsiveGrid>
 
-                {showVCDetails && vcStatus.vc_document && (
+                {showVCDetails && Boolean(vcStatus.vc_document) && (
                   <CollapsibleSection
                     title="Credential Document"
                     icon={Shield}
