+++ control-plane/web/client/src/pages/ReasonerDetailPage.tsx
@@ -9,7 +9,7 @@ import {
   Time,
   View,
 } from "../components/ui/icon-bridge";
-import { useEffect, useRef, useState } from "react";
+import { useCallback, useEffect, useRef, useState } from "react";
 import { useParams } from "react-router-dom";
 import { DIDIdentityBadge } from "../components/did/DIDDisplay";
 import { Badge } from "../components/ui/badge";
@@ -72,17 +72,11 @@ export function ReasonerDetailPage() {
   const [metrics, setMetrics] = useState<PerformanceMetrics | null>(null);
 
   // Form state
-  const [formData, setFormData] = useState<any>({});
+  const [formData, setFormData] = useState<{ input?: unknown }>({});
   const [validationErrors, setValidationErrors] = useState<string[]>([]);
   const [isExecuting, setIsExecuting] = useState(false);
 
-  useEffect(() => {
-    loadReasonerDetails();
-    loadMetrics();
-    loadHistory();
-  }, [fullReasonerId]);
-
-  const loadReasonerDetails = async () => {
+  const loadReasonerDetails = useCallback(async () => {
     if (!fullReasonerId) return;
 
     try {
@@ -102,9 +96,9 @@ export function ReasonerDetailPage() {
     } finally {
       setLoading(false);
     }
-  };
+  }, [fullReasonerId]);
 
-  const loadMetrics = async () => {
+  const loadMetrics = useCallback(async () => {
     if (!fullReasonerId) return;
     try {
       const data = await reasonersApi.getPerformanceMetrics(fullReasonerId);
@@ -112,9 +106,9 @@ export function ReasonerDetailPage() {
     } catch (err) {
       console.error("Failed to load metrics:", err);
     }
-  };
+  }, [fullReasonerId]);
 
-  const loadHistory = async () => {
+  const loadHistory = useCallback(async () => {
     if (!fullReasonerId) return;
     try {
       const data = await reasonersApi.getExecutionHistory(
@@ -126,7 +120,13 @@ export function ReasonerDetailPage() {
     } catch (err) {
       console.error("Failed to load history:", err);
     }
-  };
+  }, [fullReasonerId]);
+
+  useEffect(() => {
+    loadReasonerDetails();
+    loadMetrics();
+    loadHistory();
+  }, [loadHistory, loadMetrics, loadReasonerDetails]);
 
   const handleExecute = () => {
     if (!reasoner || !fullReasonerId || !executionQueueRef.current) return;
@@ -440,7 +440,8 @@ export function ReasonerDetailPage() {
                     </CardDescription>
                   </div>
                   {normalizeExecutionStatus(selectedExecution.status) === "succeeded" &&
-                    selectedExecution.result && (
+                    selectedExecution.result !== undefined &&
+                    selectedExecution.result !== null && (
                       <SegmentedControl
                         value={resultViewMode}
                         onValueChange={(mode) =>
@@ -455,7 +456,8 @@ export function ReasonerDetailPage() {
               </CardHeader>
               <CardContent>
                 {normalizeExecutionStatus(selectedExecution.status) === "succeeded" &&
-                selectedExecution.result ? (
+                selectedExecution.result !== undefined &&
+                selectedExecution.result !== null ? (
                   <FormattedOutput
                     data={selectedExecution.result}
                     showRaw={resultViewMode === "json"}
